@@ -6,7 +6,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.8.RELEASE</version>
+        <version>3.2.8</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.atguigu.gulimall</groupId>
@@ -16,8 +16,8 @@
 
 
     <properties>
-        <java.version>1.8</java.version>
-        <spring-cloud.version>Greenwich.SR3</spring-cloud.version>
+        <java.version>17</java.version>
+        <spring-cloud.version>2023.0.3</spring-cloud.version>
     </properties>
 
     <dependencies>
@@ -38,7 +38,7 @@
 
         <dependency>
             <groupId>com.atguigu.gulimall</groupId>
-            <artifactId>gulimall-common</artifactId>
+            <artifactId>gulimall-common-n</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
