@@ -9,6 +9,10 @@
     <description>谷粒商城-聚合服务</description>
     <packaging>pom</packaging>
 
+    <properties>
+        <java.version>17</java.version>
+    </properties>
+
     <modules>
         <module>gulimall-common</module>
         <module>gulimall-coupon</module>
@@ -19,6 +23,7 @@
         <module>gulimall-ware</module>
         <module>renren-fast</module>
 <!--        <module>renren-generator</module>-->
+        <module>gulimall-common-n</module>
     </modules>
 
     <build>
@@ -27,8 +32,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>8</source>
-                    <target>8</target>
+                    <source>17</source>
+                    <target>17</target>
                 </configuration>
             </plugin>
         </plugins>
