@@ -8,7 +8,8 @@
 
 package io.renren.common.validator.group;
 
-import javax.validation.GroupSequence;
+
+import jakarta.validation.GroupSequence;
 
 /**
  * 定义校验顺序，如果AddGroup组失败，则UpdateGroup组不会再校验
