@@ -8,8 +8,11 @@
 
 package io.renren.common.xss;
 
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
+//import jakarta.servlet.*;
+//import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+
 import java.io.IOException;
 
 /**
