@@ -10,10 +10,10 @@ package io.renren.common.validator;
 
 import io.renren.common.exception.RRException;
 import io.renren.common.utils.Constant;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
 import java.util.Set;
 
 /**
