@@ -1,16 +1,16 @@
-package io.renren;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class RenrenApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
+//package io.renren;
+//
+//import org.junit.Test;
+//import org.junit.runner.RunWith;
+//import org.springframework.boot.test.context.SpringBootTest;
+//import org.springframework.test.context.junit4.SpringRunner;
+//
+//@RunWith(SpringRunner.class)
+//@SpringBootTest
+//public class RenrenApplicationTests {
+//
+//	@Test
+//	public void contextLoads() {
+//	}
+//
+//}
